Add brew formulae for ligo
Problem: At the moment ligo lacks convenient way for installation on
macOS.

Solution: The most commonly used package manager for macOS is brew, so
we're providing a brew formula for building and installing ligo
executable. Note that in order to skip building step binary bottles have
to be built for all target macOS versions, then these bottles need to be
uploaded somewhere and listed in the 'bottle' attribute along with their
sha256 hashes in the formula.