Invoke opam directly to fix build for the new opam version
Opam 2.1.0 was released recently with minor CLI changes, and build scripts in
ligo repository are intended to be used with opam 2.0 (they use --locked=locked
flag, which needs to be changed to --locked with opam 2.1). But homebrew does
not give a a simple way to use an older opam version. So we call opam
explicitly instead of relying on the scripts, to use commands appropriate
for the new opam version.

Opam itself has OPAMCLI environment variable, which makes it support older
version of the CLI flags, but for --locked flag it does not work.